refactor(mock-repository): split mock data
